Feat: 앨범 생성 후 동기화
기존에 albumReducer를 groupReducer에 합쳐 앨범 상태 변경할 수 있도록 수정

#159
