+++ frontend/src/reducer/index.ts
@@ -5,7 +5,6 @@ import ModalReducer from "./Modal";
 import AddressReducer from "./AddressReducer";
 import Theme from "./Theme";
 import UserReducer from "./UserReducer";
-import AlbumReducer from "./AlbumReducer";
 
 const rootReducer = combineReducers({
   groupModal: GroupModalReducer,
@@ -14,7 +13,6 @@ const rootReducer = combineReducers({
   theme: Theme,
   address: AddressReducer,
   user: UserReducer,
-  album: AlbumReducer,
 });
 
 export type RootState = ReturnType<typeof rootReducer>;
